Install bp4o to /etc/profile.d and fpath
This change removes the added line in people's rc files to source bp4o scripts.

For Bash users, scripts in /etc/profile.d are automatically sourced on login.
For Zsh users, scripts in a fpath dir are loaded on demand, which means a couple lines will still need to be added to their .zshrc, but the lines are simple and inline with typical zsh configuration.